+++ packages/stream_chat_flutter/lib/src/message_list_view/message_list_view.dart
@@ -471,9 +471,6 @@ class _StreamMessageListViewState extends State<StreamMessageListView> {
 
   @override
   void dispose() {
-    if (!_upToDate) {
-      streamChannel!.reloadChannel();
-    }
     debouncedMarkRead?.cancel();
     debouncedMarkThreadRead?.cancel();
     _messageNewListener?.cancel();