+++ packages/stream_chat/lib/src/core/api/user_api.dart
@@ -91,6 +91,13 @@ class UserApi {
     return BlockedUsersResponse.fromJson(response.data);
   }
 
+  /// Requests the unread count information for the current user.
+  Future<GetUnreadCountResponse> getUnreadCount() async {
+    final response = await _client.get('/unread');
+
+    return GetUnreadCountResponse.fromJson(response.data);
+  }
+
   /// Retrieves all the active live locations of the current user.
   Future<GetActiveLiveLocationsResponse> getActiveLiveLocations() async {
     final response = await _client.get(