+++ packages/stream_chat/lib/src/core/models/message.dart
@@ -35,11 +35,7 @@ class Message extends Equatable implements ComparableFieldProvider {
     this.mentionedUsers = const [],
     this.silent = false,
     this.shadowed = false,
-    @Deprecated("Use 'reactionGroups' instead")
-    Map<String, int>? reactionCounts,
-    @Deprecated("Use 'reactionGroups' instead")
-    Map<String, int>? reactionScores,
-    Map<String, ReactionGroup>? reactionGroups,
+    this.reactionGroups,
     this.latestReactions,
     this.ownReactions,
     this.parentId,
@@ -77,11 +73,6 @@ class Message extends Equatable implements ComparableFieldProvider {
         remoteCreatedAt = createdAt,
         remoteUpdatedAt = updatedAt,
         remoteDeletedAt = deletedAt,
-        reactionGroups = _maybeGetReactionGroups(
-          reactionGroups: reactionGroups,
-          reactionCounts: reactionCounts,
-          reactionScores: reactionScores,
-        ),
         _quotedMessageId = quotedMessageId,
         _pollId = pollId;
 
@@ -129,45 +120,40 @@ class Message extends Equatable implements ComparableFieldProvider {
   @JsonKey(toJson: User.toIds)
   final List<User> mentionedUsers;
 
-  /// A map describing the count of number of every reaction.
-  @JsonKey(includeToJson: false)
-  @Deprecated("Use 'reactionGroups' instead")
-  Map<String, int>? get reactionCounts {
-    return reactionGroups?.map((type, it) => MapEntry(type, it.count));
-  }
-
-  /// A map describing the count of score of every reaction.
-  @JsonKey(includeToJson: false)
-  @Deprecated("Use 'reactionGroups' instead")
-  Map<String, int>? get reactionScores {
-    return reactionGroups?.map((type, it) => MapEntry(type, it.sumScores));
-  }
-
-  static Map<String, ReactionGroup>? _maybeGetReactionGroups({
-    Map<String, ReactionGroup>? reactionGroups,
-    Map<String, int>? reactionCounts,
-    Map<String, int>? reactionScores,
-  }) {
+  static Object? _reactionGroupsReadValue(
+    Map<Object?, Object?> json,
+    String key,
+  ) {
+    final reactionGroups = json[key] as Map<String, dynamic>?;
     if (reactionGroups != null) return reactionGroups;
+
+    final reactionCounts = json['reaction_counts'] as Map<String, dynamic>?;
+    final reactionScores = json['reaction_scores'] as Map<String, dynamic>?;
     if (reactionCounts == null && reactionScores == null) return null;
 
     final reactionTypes = {...?reactionCounts?.keys, ...?reactionScores?.keys};
     if (reactionTypes.isEmpty) return null;
 
-    final groups = <String, ReactionGroup>{};
+    final groups = <String, dynamic>{};
     for (final type in reactionTypes) {
       final count = reactionCounts?[type] ?? 0;
       final sumScores = reactionScores?[type] ?? 0;
 
       if (count == 0 || sumScores == 0) continue;
-      groups[type] = ReactionGroup(count: count, sumScores: sumScores);
+      final now = DateTime.timestamp();
+      groups[type] = {
+        'count': count,
+        'sum_scores': sumScores,
+        'first_reaction_at': now.toIso8601String(),
+        'last_reaction_at': now.toIso8601String(),
+      };
     }
 
     return groups;
   }
 
   /// A map of reaction types and their corresponding reaction groups.
-  @JsonKey(includeToJson: false)
+  @JsonKey(includeToJson: false, readValue: _reactionGroupsReadValue)
   final Map<String, ReactionGroup>? reactionGroups;
 
   /// The latest reactions to the message created by any user.
@@ -401,10 +387,6 @@ class Message extends Equatable implements ComparableFieldProvider {
     List<User>? mentionedUsers,
     bool? silent,
     bool? shadowed,
-    @Deprecated("Use 'reactionGroups' instead")
-    Map<String, int>? reactionCounts,
-    @Deprecated("Use 'reactionGroups' instead")
-    Map<String, int>? reactionScores,
     Map<String, ReactionGroup>? reactionGroups,
     List<Reaction>? latestReactions,
     List<Reaction>? ownReactions,
@@ -477,12 +459,7 @@ class Message extends Equatable implements ComparableFieldProvider {
       mentionedUsers: mentionedUsers ?? this.mentionedUsers,
       silent: silent ?? this.silent,
       shadowed: shadowed ?? this.shadowed,
-      reactionGroups: _maybeGetReactionGroups(
-            reactionGroups: reactionGroups,
-            reactionCounts: reactionCounts,
-            reactionScores: reactionScores,
-          ) ??
-          this.reactionGroups,
+      reactionGroups: reactionGroups ?? this.reactionGroups,
       latestReactions: latestReactions ?? this.latestReactions,
       ownReactions: ownReactions ?? this.ownReactions,
       parentId: parentId ?? this.parentId,