+++ packages/stream_chat_flutter/test/src/mocks.dart
@@ -68,6 +68,7 @@ class MockChannelState extends Mock implements ChannelClientState {
     when(() => typingEvents).thenReturn({});
     when(() => typingEventsStream).thenAnswer((_) => Stream.value({}));
     when(() => unreadCount).thenReturn(0);
+    when(() => isUpToDate).thenReturn(true);
     when(() => read).thenReturn([]);
   }
 }